+++ NEWS.md
@@ -3,4 +3,7 @@
 * Added a `NEWS.md` file to track changes to the package.
 * Build supporting documents for package.
 * Add initial functions from homeworks.
-
+  - `f_a_h()` - generate vector centered around a, bounded by h.
+  - `slope_x1_x2()` - calculate the rise over run for a point of f(x)
+  - `limit_at_a()` - calculate the limit at point a for f(x) from the slopes.
+  - `dy_dx()` - calculate the points on the dy/dx, or derivative, for a function f(x).